--- lama-template/src/data/UserDatabase.ts
@@ -1,42 +1,0 @@
-import { BaseDatabase } from "./BaseDatabase";
-import { User } from "../model/User";
-
-export class UserDatabase extends BaseDatabase {
-  protected tableName: string = "lama_users";
-
-  private toModel(dbModel?: any): User | undefined {
-    return (
-       dbModel &&
-       new User(
-          dbModel.id,
-          dbModel.name,
-          dbModel.email,
-          dbModel.password,
-          dbModel.role
-       )
-    );
- }
-
-  public async createUser(user: User): Promise<void> {
-    try {
-      await this.getConnection().insert(user).into(this.tableName);
-    } catch (error) {
-      throw new Error(error.sqlMessage || error.message);
-    }
-  }
-
-  public async getUserByEmail(email: string): Promise<User | undefined> {
-    try {
-      const result = await this.getConnection()
-        .select("*")
-        .from(this.tableName)
-        .where({ email });
-
-      return this.toModel(result[0]);
-    } catch (error) {
-      throw new Error(error.sqlMessage || error.message);
-    }
-  }
-}
-
-export default new UserDatabase();
