--- lama-template/src/error/BaseError.ts
@@ -1,8 +1,0 @@
-export class BaseError extends Error {
-  constructor(
-     public statusCode: number,
-     message: string
-  ) {
-     super(message);
-  }
-}
