--- lama-template/src/routes/BandRouter.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import bandController from "../controller/BandController";
-
-
-export const bandRouter = express.Router();
-
-bandRouter.post("/register", bandController.create);
-bandRouter.post("/login", );
