--- lama-template/src/routes/userRouter.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import userController from "../controller/UserController";
-
-
-export const userRouter = express.Router();
-
-userRouter.post("/signup", userController.signup);
-userRouter.post("/login", userController.login);
