--- lama-template/src/services/IdGenerator.ts
@@ -1,10 +1,0 @@
-import { v4 } from "uuid";
-
-export class IdGenerator {
-    public generate(): string {
-      return v4();
-    }
-  }
-  
-  export default new IdGenerator()
-
