+++ src/error/BaseError.ts
@@ -1,0 +1,5 @@
+export class BaseError extends Error {
+  constructor(public statusCode: number, message: string) {
+    super(message);
+  }
+}
