+++ src/index.ts
@@ -1,0 +1,23 @@
+import dotenv from "dotenv";
+import { AddressInfo } from "net";
+import express from "express";
+import { userRouter } from "./routes/UserRouter";
+import { bandRouter } from "./routes/BandRouter";
+import { showRouter } from "./routes/ShowRouter";
+dotenv.config();
+const app = express();
+
+app.use(express.json());
+
+app.use("/user", userRouter);
+app.use("/band", bandRouter);
+app.use("/show", showRouter);
+
+const server = app.listen(3000, () => {
+  if (server) {
+    const address = server.address() as AddressInfo;
+    console.log(`Servidor rodando em http://localhost:${address.port}`);
+  } else {
+    console.error(`Falha ao rodar o servidor.`);
+  }
+});
