+++ src/routes/BandRouter.ts
@@ -1,0 +1,7 @@
+import express from "express";
+import bandController from "../controller/BandController";
+
+export const bandRouter = express.Router();
+
+bandRouter.post("/register", bandController.create);
+bandRouter.get("", bandController.searchByIdOrName);
