+++ src/routes/UserRouter.ts
@@ -1,0 +1,7 @@
+import express from "express";
+import userController from "../controller/UserController";
+
+export const userRouter = express.Router();
+
+userRouter.post("/signup", userController.signup);
+userRouter.post("/login", userController.login);
