+++ tests/mocks/authenticatorMock.ts
