+++ tests/mocks/hashGeneratorMock.ts
